chore: move global and repo to root
