+++ config.yaml
@@ -6,6 +6,7 @@ WebServer:
 
 Redis:
   Db: 0
+  QueueDb: 0
   Address: '127.0.0.1:6379'
   Password: ''
