+++ internal/api/debug/handler.go
@@ -1,7 +1,7 @@
 package debug
 
 import (
-	"github.com/WHUPRJ/woj-server/internal/global"
+	"github.com/WHUPRJ/woj-server/global"
 	"github.com/gin-gonic/gin"
 	"go.uber.org/zap"
 )
