+++ internal/api/user/profile.go
@@ -1,8 +1,8 @@
 package user
 
 import (
+	"github.com/WHUPRJ/woj-server/global"
 	"github.com/WHUPRJ/woj-server/internal/e"
-	"github.com/WHUPRJ/woj-server/internal/global"
 	"github.com/WHUPRJ/woj-server/model"
 	"github.com/gin-gonic/gin"
 )
