+++ internal/app/app.go
@@ -3,11 +3,11 @@ package app
 import (
 	"context"
 	"fmt"
-	"github.com/WHUPRJ/woj-server/internal/global"
-	"github.com/WHUPRJ/woj-server/internal/repo/postgresql"
-	"github.com/WHUPRJ/woj-server/internal/repo/redis"
+	"github.com/WHUPRJ/woj-server/global"
 	"github.com/WHUPRJ/woj-server/internal/router"
 	"github.com/WHUPRJ/woj-server/internal/service/jwt"
+	"github.com/WHUPRJ/woj-server/repo/postgresql"
+	"github.com/WHUPRJ/woj-server/repo/redis"
 	"go.uber.org/zap"
 	"net/http"
 	"os"
