+++ internal/service/jwt/service.go
@@ -1,7 +1,7 @@
 package jwt
 
 import (
-	"github.com/WHUPRJ/woj-server/internal/global"
+	"github.com/WHUPRJ/woj-server/global"
 	"github.com/go-redis/redis/v8"
 	"go.uber.org/zap"
 )
