+++ internal/service/problem/service.go
@@ -1,8 +1,8 @@
 package problem
 
 import (
+	"github.com/WHUPRJ/woj-server/global"
 	"github.com/WHUPRJ/woj-server/internal/e"
-	"github.com/WHUPRJ/woj-server/internal/global"
 	"github.com/WHUPRJ/woj-server/model"
 	"go.uber.org/zap"
 	"gorm.io/gorm"
