+++ internal/service/task/service.go
@@ -1,8 +1,8 @@
 package task
 
 import (
+	"github.com/WHUPRJ/woj-server/global"
 	"github.com/WHUPRJ/woj-server/internal/e"
-	"github.com/WHUPRJ/woj-server/internal/global"
 	"github.com/WHUPRJ/woj-server/model"
 	"github.com/hibiken/asynq"
 	"go.uber.org/zap"
@@ -11,7 +11,9 @@ import (
 var _ Service = (*service)(nil)
 
 type Service interface {
-	NewJudge(submission model.Submission) e.Status
+	NewJudge(submission model.Submission) (string, e.Status)
+	PushProblem(id uint, file string) (string, e.Status)
+	GetTaskInfo(id string) (*asynq.TaskInfo, e.Status)
 	submit(typename string, payload []byte) (*asynq.TaskInfo, e.Status)
 }
 
@@ -25,7 +27,7 @@ func NewService(g *global.Global) Service {
 	redisOpt := asynq.RedisClientOpt{
 		Addr:     g.Conf.Redis.Address,
 		Password: g.Conf.Redis.Password,
-		DB:       g.Conf.Redis.Db,
+		DB:       g.Conf.Redis.QueueDb,
 	}
 	return &service{
 		log:       g.Log,
