+++ internal/service/user/service.go
@@ -1,8 +1,8 @@
 package user
 
 import (
+	"github.com/WHUPRJ/woj-server/global"
 	"github.com/WHUPRJ/woj-server/internal/e"
-	"github.com/WHUPRJ/woj-server/internal/global"
 	"github.com/WHUPRJ/woj-server/model"
 	"github.com/go-redis/redis/v8"
 	"go.uber.org/zap"
